+++ scrappers/utils/mixins.py
@@ -1,0 +1,5 @@
+class Mixin:
+    def clean_text(self, text):
+        """Takeout trailing spaces from a given text
+        """
+        return text.strip()